pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "hikari>=2.1.0",
    "hikari-lightbulb>=2.3.5.post1",
    "python-dotenv>=1.0.1",
]
